Implement batch saving in hydra_pg
