+++ db/index.ts
@@ -1,45 +1,28 @@
 import { createClient } from "@libsql/client";
+import seed from "./seed.sql?raw";
 
+const DATABASE_URL = process.env.DATABASE_URL || "http://0.0.0.0:8080";
 const client = createClient({
-    url: "http://127.0.0.1:8080",
+    url: DATABASE_URL,
 });
 
-// Create a table
-await client.execute(
-    "CREATE TABLE IF NOT EXISTS users (id INTEGER PRIMARY KEY, name TEXT, email TEXT)"
-);
+// Delete comments and empty lines "--"
+const noComments = seed.replace(/--.*/g, "");
+const noEmptyLines = noComments.replace(/\n\s*\n/g, "");
 
-// Insert a row
+const statements = noEmptyLines.split(";").map(s => s.trim()).filter(s => s.length > 0);
 
-function getRandomString(length: number) {
-    const characters =
-        "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789";
-    let result = "";
-    const charactersLength = characters.length;
-    for (let i = 0; i < length; i++) {
-        result += characters.charAt(
-            Math.floor(Math.random() * charactersLength)
-        );
+
+for (const statement of statements) {
+    try {
+        const r = await client.execute(statement);
+        console.log(statement.split("\n")[0] + "  ...OK")
+    } catch (e) {
+        console.error("Statement failed:")
+        console.log("--------------------------------")
+        console.log(statement)
+        console.log("--------------------------------")
+        console.error(e)
+        console.log("--------------------------------")
     }
-    return result;
-}
-
-for (let index = 0; index < 1000; index++) {
-    const randomName = getRandomString(10);
-    const randomEmail = `${getRandomString(10)}@example.com`;
-
-    const insertResult = await client.execute({
-        sql: "INSERT INTO users (name, email) VALUES (?, ?) RETURNING id",
-        args: [randomName, randomEmail],
-    });
-    // Query the row
-    const userId = insertResult.rows[0].id;
-    const result = await client.execute({
-        sql: "SELECT * FROM users WHERE id = ?",
-        args: [userId],
-    });
-
-    // Print the result
-    const row = result.rows[0];
-    console.log(`Name: ${row.name}, Email: ${row.email}`);
 }
