+++ docker-compose.yml
@@ -2,55 +2,22 @@ name: rgs
 
 services:
     db:
-        image: timescale/timescaledb:latest-pg16
+        image: ghcr.io/tursodatabase/libsql-server:latest
         environment:
-            POSTGRES_USER: postgres
-            POSTGRES_PASSWORD: postgres
-
+            SQLD_NODE: standalone
         volumes:
-            - ./data:/var/lib/postgresql/data
+            - ./db/data:/var/lib/sqld
         ports:
-            - "5432:5432"
-        healthcheck:
-            test: ["CMD", "pg_isready", "-U", "postgres", "-d", "postgres"]
-            interval: 1s
-            timeout: 1s
-            retries: 25
-
+            - "8080:8080"
     migration:
-        build:
-            context: ./db
-        depends_on:
-            db:
-                condition: service_healthy
+        image: oven/bun:1
         volumes:
-            - ./db/hasura_dump:/app/hasura_dump
-        command: ["npm", "run", "bootstrap"]
+            - ./db:/app
+        working_dir: /app
+        command: [ "bun", "index.ts" ]
         environment:
-            DATABASE_URL: postgres://postgres:postgres@db:5432/postgres
-
-    grafana:
-        container_name: grafana
-        image: grafana/grafana-enterprise
-        user: "root:root"
-        environment:
-            GF_DATABASE_TYPE: postgres
-            GF_DATABASE_HOST: db
-            GF_DATABASE_NAME: postgres
-            GF_DATABASE_USER: postgres
-            GF_DATABASE_PASSWORD: postgres
-            GF_DATABASE_SSL_MODE: disable
-            GF_DASHBOARDS_MIN_REFRESH_INTERVAL: 200ms
-            GF_AUTH_ANONYMOUS_ENABLED: true
-            GF_SECURITY_ALLOW_EMBEDDING: true
-        restart: unless-stopped
-        depends_on:
-            - db
-        ports:
-            - 0.0.0.0:5657:3000
-        volumes:
-            - ./grafana_storage:/var/lib/grafana
-
+            DATABASE_URL: http://db:8080
+        restart: on-failure:3
     # hydrand:
     #   build:
     #     context: .
@@ -72,7 +39,7 @@ services:
     #     context: ./hydra_pg
     #   dockerfile: Dockerfile
     #   command: >
-    #     --db-url postgres://postgres:postgres@db:5432/postgres
+    #     --db-url libsql://db:8080
     #     --address 0.0.0.0
     #     --port 5656
     #   ports:
