+++ tile_provider/Dockerfile
@@ -1,4 +1,4 @@
-FROM oven/bun:1 AS base
+FROM oven/bun AS base
 WORKDIR /usr/src/app
 
 FROM base AS install
@@ -11,18 +11,13 @@ COPY package.json bun.lockb /temp/prod/
 RUN cd /temp/prod && bun install --frozen-lockfile --production
 
 FROM base AS prerelease
-COPY --from=install /temp/dev/node_modules node_modules
+COPY --from=install /temp/prod/node_modules node_modules
 COPY . .
 
 RUN bun test
-RUN bun build /usr/src/app/index.ts --compile --outfile tile_provider
 
 FROM base AS release
-COPY --from=prerelease /usr/src/app/tile_provider /usr/local/bin/tile_provider
+COPY --from=prerelease . .
 
-RUN mkdir /var/lib/tile_provider
-RUN chown bun:bun /var/lib/tile_provider
-
-USER bun
-EXPOSE 8080/tcp
-ENTRYPOINT [ "/usr/local/bin/tile_provider", "--database", "/var/lib/tile_provider/tiles.db", "--port", "8080" ]
+EXPOSE 6565/tcp
+ENTRYPOINT [  "bun", "run", "/usr/src/app/index.ts", "--database", "/var/lib/tile_provider/tiles.db", "--port", "6565" ]
