+++ tile_provider/public/index.html
@@ -9,6 +9,6 @@
     </style>
   </head>
   <body>
-    <script type="module" src="/src/index.ts"></script>
+    <script type="module" src="../client/index.ts"></script>
   </body>
 </html>
